# openapi.yaml
openapi: 3.0.3
info:
  title: Secure Legal DMS API
  version: 0.1.0
  description: >
    Backend contract for the SIH 2026 secure document management system.
    This file is the source of truth. Do not change an endpoint without
    telling the whole team in writing.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - sessionToken: []
paths:
...
  /cases:
    get:
      tags: [cases]
      summary: List cases the caller is assigned to.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Assigned cases only, never all cases
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Case'
                  total:
                    type: integer
    post:
      tags: [cases]
      summary: >
        Open a case (case.create). Station and creator come from the
        session; the creator is assigned in the same transaction.
        Audited as case_create.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [case_number, title]
              properties:
                case_number:
                  type: string
                title:
                  type: string
                sensitivity:
                  $ref: '#/components/schemas/Sensitivity'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Case number already exists
  /cases/{case_id}:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    get:
      tags: [cases]
      summary: Get one case, with the officers currently assigned.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [cases]
      summary: >
        Change title, status or sensitivity (case.update). Audited as
        case_update with every changed field as {from, to}.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                status:
                  type: string
                  enum: [open, under_investigation, charge_sheeted, closed]
                sensitivity:
                  $ref: '#/components/schemas/Sensitivity'
      responses:
        '200':
          description: Updated (or unchanged)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          description: Nothing to change, or an invalid value
        '403':
          description: >
            Your rank does not permit this, or the change would lower the case's
            sensitivity - that only ever goes up, since lowering it switches off
            the redaction applied to evidence already released
  /cases/{case_id}/documents:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    get:
      tags: [documents]
      summary: List documents in a case.
      parameters:
        - name: doc_type
          in: query
          schema:
            $ref: '#/components/schemas/DocType'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '403':
          $ref: '#/components/responses/Forbidden'
  /cases/{case_id}/protected-identities:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    get:
      tags: [cases]
      summary: >
        Identities redaction removes from every export on this case.
        Readable only by officers assigned to the case. Never exported.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProtectedIdentity'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [cases]
      summary: >
        Flag an identity as protected. Deciding who counts is an
        investigative judgement, so this needs case.assign.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: string
                kind:
                  type: string
                  enum: [name, address, phone, relationship]
                  default: name
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtectedIdentity'
        '400':
          description: value is required
        '403':
          $ref: '#/components/responses/Forbidden'
  /cases/{case_id}/assignments:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [cases]
      summary: Put an officer on the case (case.assign). Audited as assign.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Assigned
        '404':
          description: No active officer with that id
        '409':
          description: Already on the case
  /cases/{case_id}/assignments/{user_id}:
    parameters:
      - $ref: '#/components/parameters/CaseId'
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [cases]
      summary: >
        Take an officer off the case (case.assign). Audited as unassign.
        You cannot remove yourself, so someone able to assign always remains.
      responses:
        '204':
          description: Removed
        '400':
          description: Tried to remove yourself
        '404':
          description: That officer is not on this case
  /cases/{case_id}/audit:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    get:
      tags: [audit]
      summary: >
        Every event on this case, oldest first - creation, edits,
        assignments, evidence filed, viewed, versioned, verified,
        released, and refused attempts - each with the acting officer.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  chain_intact:
                    type: boolean
                  broken_at:
                    type: integer
                    nullable: true
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEntry'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: Opaque session token from /auth/mfa/verify
  parameters:
    CaseId:
      name: case_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 25
        maximum: 100
  schemas:
    DocType:
      type: string
      enum: [fir, statement, forensic_report, charge_sheet, court_filing, notice, other]
    Rank:
      type: string
      enum: [constable, head_constable, sub_inspector, inspector, dcp, prosecutor, judge, forensic_analyst]
    Sensitivity:
      type: string
      description: >
        "protected" triggers mandatory redaction on every export path,
        for offences covered by Section 72 BNS.
      enum: [normal, restricted, protected]
...
    ProtectedIdentity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        value:
          type: string
        kind:
          type: string
          enum: [name, address, phone, relationship]
        added_at:
          type: string
          format: date-time
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        service_number:
          type: string
        rank:
          $ref: '#/components/schemas/Rank'
        station:
          type: string
    Case:
      type: object
      properties:
        id:
          type: string
          format: uuid
        case_number:
          type: string
          example: FIR/0142/2026
        title:
          type: string
        sensitivity:
          $ref: '#/components/schemas/Sensitivity'
        status:
          type: string
          enum: [open, under_investigation, charge_sheeted, closed]
        created_at:
          type: string
          format: date-time
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        case_id:
          type: string
          format: uuid
        evidence_number:
          type: string
          description: Assigned by the database, unique within the case, never changes
          example: STM-002
        title:
          type: string
        doc_type:
          $ref: '#/components/schemas/DocType'
        current_version:
          type: integer
        created_at:
          type: string
          format: date-time
...
    AuditEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        actor:
          $ref: '#/components/schemas/User'
        action:
          type: string
          enum: [upload, view, download, export_redacted, new_version, verify, search,
                 share, access_denied, login, case_create, case_update, assign,
                 unassign, identity_add, delete_attempt]
        case_id:
          type: string
          format: uuid
          nullable: true
          description: Null only for login, search and pre-case denials (enforced by the database)
        document_id:
          type: string
          format: uuid
        evidence_number:
          type: string
          nullable: true
        detail:
          type: object
          nullable: true
          description: Action-specific; case_update carries changes as {field - {from, to}}
        occurred_at:
          type: string
          format: date-time
        prev_hash:
          type: string
        entry_hash:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string
          example: forbidden
        message:
          type: string
          example: You are not assigned to this case.
  responses:
...
    Forbidden:
      description: >
        Logged in but not permitted. Always logged as access_denied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: >
        Returned instead of 403 where confirming existence would itself
        leak information about a protected case.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'